Function: allow `keepNames` and `define` esbuild config options
* Allow keepNames esbuild config option

This commit allows the keepNames option in the esbuild config. This is done by creating an array of allowed options for the esbuild config that can be added to easily. This is also reflected in the error message for using a non-allowed option.

* Rename esbuildconfig test directory

This commit renames the directory to reflect the new options that are allowed.

* Revert "Rename esbuildconfig test directory"

This reverts commit b06fbaf6dd1e824b712dbb39cda6b510938edc1c.

* Revert "Allow keepNames esbuild config option"

This reverts commit 5c72e906b39323dfbb7ec63a1388ed53939cc5aa.

* Support inline config

* Update docs

Co-authored-by: Frank <wangfanjie@gmail.com>