Clarify that additional pkg needs to be installed
As a beginner working through this page of the docs I spent a bit of time struggling to make the `@serverless-stack/node/config` import work, not realizing that I needed to manually install an additional npm package. Other users may assume, like me, that create-sst would install all needed packages automatically. I think it would be useful to add a bit of clarification about that point on this page to remove friction from the learning process. I've made changes I think would be best, but the exact wording and positioning are of course up for debate.