apply tags via TagManager instead of tags aspect
This updates the App construct to apply tags via TagManager instead of the tags aspect. What that means is that it will apply the tags via the native Stack mechanism, e.g stackTags instead of adding tags to all cloudformation resources individually. This makes the resulting stack less verbose and doesn't result in things getting tagged that aren't supported. For example, OAM Sink breaks if you apply tags since it's schema is strict.