feat: add a single lifecyle rule to stop triggering rule S3.13 in Security Hub
This is a low level warning generated when enabling AWS Foundational
Security Best Practices v1.0.0 in Security Hub. This is a very common
one to enable. Just adding an almost empty life-cyle rule will stop
this warning.

It might be good to also add a rule to transition old objects to
glacier, or remove them altogether.