Delete side effect only @aws-sdk/types import
"Bare" imports that don't name what they are importing are never elided
by Typescript and end up in the compiled JS. That doesn't make any sense
for this package since it only has types.

I don't know why this was originally added, but it still typechecks
without it now so I don't think it's still necessary.