fix: Handler wrapper should honor path for nested handlers (#469)
The existing code added a `dir` property to the returned `wrapper` object. This property was never read or used. The `wrapper` object is used as input to an archiver, which supports full paths as part of the `name` key. 

This PR removes the `dir` key for the wrapper and just prepends the path to the `name` key.