Finalized sst.Linkable api (#715)
This is a complete reworking of the link system so if you were using some of the internals things may break. You will get explicit errors if you try to use the old system. We are still working on docs for the new system but you can see an example in examples/aws-linkable